Rename binary_state_controller->binary_controller
