+++ binary_controller/include/binary_controller/binary_state_interface.h
