--- binary_state_controller/binary_state_controller_plugins.xml
@@ -1,7 +1,0 @@
-<library path="lib/libbinary_state_controller">
-  <class name="binary_state_controller/BinaryCommandController" type="binary_state_controller::BinaryCommandController" base_class_type="controller_interface::ControllerBase">
-    <description>
-      Command a binary device
-    </description>
-  </class>
-</library>
